Refactored workoutHistoryTableRow , moved all workout components to their own folder
